+++ packages/adminui-vue/forms/aclRlsEntry-fm.vue
@@ -1,13 +1,13 @@
 <template>
   <div class="u-form-layout">
-    <u-toolbar />
+    <u-toolbar :show-dropdown="false" />
 
     <u-form-container
       v-loading="loading"
       label-position="top"
     >
       <u-form-row
-        label="aclRlsEntry_form.entity"
+        label="aclRlsInfo.subject"
         class="u-acl-rls-input-overflow"
       >
         <el-select
@@ -59,31 +59,44 @@
 const { Form } = require('@unitybase/adminui-vue')
 const { Repository, connection } = require('@unitybase/ub-pub')
 const { mapMutations, mapActions, mapGetters } = require('vuex')
+const { UBDomain } = require('@unitybase/cs-shared')
 
 const ACL_RLS_COLLECTION = 'aclRlsEntries'
 
-module.exports.mount = cfg => {
-  const { instanceID } = cfg.props
+async function loadOrdersByEnumGroup (enumGroup) {
+  const entries = await Repository('ubm_enum')
+    .attrs('sortOrder', 'code', 'eGroup')
+    .where('eGroup', '=', enumGroup)
+    .select()
 
-  Form({
-    ...cfg,
-    props: {
-      ...cfg.props,
-      parentContext: { instanceID }
-    }
-  })
-    .store()
+  return Object.fromEntries(entries.map(({ code, sortOrder }) => [code, sortOrder]))
+}
+
+/**
+ * The form loads entry of master entity for which user configures aclRls access.
+ * This entry NEVER be changed and updated. AclRls entries are loaded and managed
+ * as a collection since the user can add several items from the one form
+ */
+module.exports.mount = cfg => {
+  const { aclEntityName, aclAttributes, instanceID } = cfg.props
+
+  Form(cfg)
+    .store({
+      getters: {
+        canDelete () {
+          return false
+        }
+      }
+    })
     .processing({
+      masterFieldList: ['ID'], // not load all attributes to not overload server
+
       collections: {
         [ACL_RLS_COLLECTION] () {
-          return Repository(cfg.entity)
-            .attrs('ID', 'instanceID', ...cfg.props.aclAttributes.map(attr => attr.code))
+          return Repository(aclEntityName)
+            .attrs('ID', 'instanceID', aclAttributes.map(attr => attr.code))
             .where('instanceID', '=', instanceID)
         }
-      },
-
-      inited ({ commit }) {
-        commit('IS_NEW', false)
       }
     })
     .mount()
@@ -93,6 +106,7 @@ export default {
   name: 'AclRlsEntry',
 
   props: {
+    /** @type {UBEntityAttribute[]} */
     aclAttributes: {
       type: Array,
       required: true
@@ -107,30 +121,22 @@ export default {
   data () {
     return {
       selectedEntity: null,
-      aclRlsEntries: null
+      aclRlsEntries: null,
+      mappedEntitiesForAclAttrs: {}
     }
   },
 
   computed: {
-    ...mapGetters(['loading']),
+    ...mapGetters([
+      'loading'
+    ]),
 
     collectionItems () {
       return this.$store.state.collections[ACL_RLS_COLLECTION].items.map(item => item.data)
     },
 
-    mappedEntitiesForAclAttrs () {
-      return Object.fromEntries(
-        this.aclAttributes.map(attrDef => {
-          const mappedEntities = connection.domain
-            .filterEntities(entityDef => entityDef.mixin('unity')?.entity === attrDef.associatedEntity)
-            .map(entityDef => entityDef.code)
-          return [attrDef.associatedEntity, mappedEntities]
-        })
-      )
-    },
-
     mappedEntitiesBySelectedEntity () {
-      return this.mappedEntitiesForAclAttrs[this.selectedEntity]
+      return this.mappedEntitiesForAclAttrs[this.selectedEntity] ?? []
     },
 
     aclRlsEntriesKey () {
@@ -148,21 +154,73 @@ export default {
     }
   },
 
-  mounted () {
-    this.sortMappedEntites()
+  async mounted () {
+    await this.loadMappedEntitesForUnityOnes()
+
+    this.selectedEntity = this.aclAttributes[0].associatedEntity
   },
 
   methods: {
     ...mapMutations([
-      'DELETE_COLLECTION_ITEM'
+      'DELETE_COLLECTION_ITEM_WITHOUT_TRACKING'
     ]),
 
     ...mapActions([
       'addCollectionItemWithoutDefaultValues'
     ]),
 
-    sortMappedEntites () {
+    async loadMappedEntitesForUnityOnes () {
+      const mappedEntitiesDict = {}
 
+      for (const { associatedEntity } of this.aclAttributes) {
+        const mappedEntities = connection.domain
+          .filterEntities(entityDef => entityDef.mixin('unity')?.entity === associatedEntity)
+          .map(entityDef => entityDef.code)
+        await this.sortMappedEntities(mappedEntities, associatedEntity)
+
+        mappedEntitiesDict[associatedEntity] = mappedEntities
+      }
+
+      this.mappedEntitiesForAclAttrs = mappedEntitiesDict
+    },
+
+    async sortMappedEntities (entities, associatedEntity) {
+      const getUnityDefaults = entity => connection.domain.get(entity).mixin('unity').defaults
+      const getUnityDefaultsKey = entity => Object.keys(getUnityDefaults(entity))
+      const arraysIntersaction = (a1, a2) => a1.filter(key => a2.includes(key))
+
+      let commonAttributes = getUnityDefaultsKey(entities[0])
+      for (let i = 1; i < entities.length; i++) {
+        commonAttributes = arraysIntersaction(commonAttributes, getUnityDefaultsKey(entities[i]))
+      }
+
+      if (commonAttributes.length !== 1) {
+        throw new Error(
+          'You should define only one default value that will be mapped to the unity entity as it defines an order for displaying in UI'
+        )
+      }
+
+      // set order of controls for unity entities based on values of ONE common attribute
+      const [orderAttrName] = commonAttributes
+
+      const getComparator = (reflect = v => v) => (e1, e2) => {
+        const order1 = reflect(getUnityDefaults(e1)[orderAttrName])
+        const order2 = reflect(getUnityDefaults(e2)[orderAttrName])
+        if (!order1 || !order2) {
+          return 0
+        }
+        return order1 > order2 ? 1 : -1
+      }
+
+      const orderAttrDef = connection.domain.get(associatedEntity).attributes[orderAttrName]
+      const isEnumDataType = orderAttrDef.dataType === UBDomain.ubDataTypes.Enum
+
+      if (isEnumDataType) { // compare by sortOrder of enums
+        const entries = await loadOrdersByEnumGroup(orderAttrDef.enumGroup)
+        entities.sort(getComparator(code => entries[code]))
+      } else { // compare by chars
+        entities.sort(getComparator())
+      }
     },
 
     resetSelectedItems () {
@@ -177,6 +235,13 @@ export default {
       )
     },
 
+    /**
+     * @param {object} params
+     * @param {object[]} [params.source] array of aclRls entries
+     * @param {object[]} params.unityEntity associated entity for acl attribute, values of which we want to load
+     * @param {object[]} [params.entity] subentity if the associtaed entity is a unity for several ones
+     * @returns {number[]}
+     */
     pickAclAttributeValues ({
       source = this.aclRlsEntries,
       unityEntity,
@@ -214,8 +279,9 @@ export default {
       const promises = []
 
       for (const aclAttr of this.aclAttributes) {
-        const currentIds = this.pickAclAttributeValues({ source: JSON.parse(value), unityEntity: aclAttr.associatedEntity })
-        const prevIds = this.pickAclAttributeValues({ source: JSON.parse(prevValue), unityEntity: aclAttr.associatedEntity })
+        const unityEntity = aclAttr.associatedEntity
+        const currentIds = this.pickAclAttributeValues({ source: JSON.parse(value), unityEntity })
+        const prevIds = this.pickAclAttributeValues({ source: JSON.parse(prevValue), unityEntity })
 
         const newIds = currentIds.filter(id => !prevIds.includes(id))
         for (const valueID of newIds) {
@@ -230,10 +296,12 @@ export default {
           )
         }
 
-        const removedIds = prevIds.filter(id => currentIds.includes(id))
+        const removedIds = prevIds.filter(id => !currentIds.includes(id))
+
+        // iterate from the end to not mix indexes if we should delete several items
         for (const [index, item] of [...this.collectionItems.entries()].reverse()) {
           if (removedIds.includes(item[aclAttr.code])) {
-            this.DELETE_COLLECTION_ITEM({
+            this.DELETE_COLLECTION_ITEM_WITHOUT_TRACKING({
               collection: ACL_RLS_COLLECTION,
               index
             })